Rename conv factory.
